Bluetooth: audio: has: Move HAS Client API to internal
This moves HAS Client API to internal.
The user shall use HAP HARC API instead.
