+++ subsys/bluetooth/audio/has_internal.h
@@ -158,3 +158,273 @@ static inline bool bt_has_features_check_preset_write_supported(enum bt_has_feat
 {
 	return (features & BT_HAS_FEAT_DYNAMIC_PRESETS) > 0;
 }
+
+/** @brief Opaque Hearing Access Service Client instance. */
+struct bt_has_client;
+
+/** @brief Hearing Access Service Client callback structure. */
+struct bt_has_client_cb {
+	/**
+	 * @brief Service has been connected.
+	 *
+	 * This callback notifies the application of a new service connection.
+	 * In case the err parameter is non-zero it means that the connection establishment failed.
+	 *
+	 * @param client Client instance.
+	 * @param err 0 in case of success or negative errno or ATT Error code in case of error.
+	 */
+	void (*connected)(struct bt_has_client *client, int err);
+
+	/**
+	 * @brief Service has been disconnected.
+	 *
+	 * This callback notifies the application that the service has been disconnected.
+	 *
+	 * @param client Client instance.
+	 * @param unbound True if service binding has been removed.
+	 */
+	void (*disconnected)(struct bt_has_client *client);
+
+	/**
+	 * @brief Client instance unbound.
+	 *
+	 * This callback notifies the application that the client instance has been free'd.
+	 *
+	 * @param client Client instance.
+	 * @param err 0 in case of success or negative errno or ATT Error code in case of error.
+	 */
+	void (*unbound)(struct bt_has_client *client, int err);
+
+	/**
+	 * @brief Callback function for Hearing Access Service active preset changes.
+	 *
+	 * Optional callback called when the active preset is changed by the remote server when the
+	 * preset switch procedure is complete. The callback must be set to receive active preset
+	 * changes and enable support for switching presets. If the callback is not set, the Active
+	 * Index and Control Point characteristics will not be discovered by
+	 * @ref bt_has_client_discover.
+	 *
+	 * @param client Client instance.
+	 * @param index Active preset index.
+	 */
+	void (*preset_switch)(struct bt_has_client *client, uint8_t index);
+
+	/**
+	 * @brief Callback function for presets read operation.
+	 *
+	 * The callback is called when the preset read response is sent by the remote server.
+	 * The record object as well as its members are temporary and must be copied to in order
+	 * to cache its information.
+	 *
+	 * @param client Client instance.
+	 * @param record Preset record or NULL on errors.
+	 * @param is_last True if Read Presets operation can be considered concluded.
+	 */
+	void (*preset_read_rsp)(struct bt_has_client *client,
+				const struct bt_has_preset_record *record, bool is_last);
+
+	/**
+	 * @brief Callback function for preset update notifications.
+	 *
+	 * The callback is called when the preset record update is notified by the remote server.
+	 * The record object as well as its objects are temporary and must be copied to in order
+	 * to cache its information.
+	 *
+	 * @param client Client instance.
+	 * @param index_prev Index of the previous preset in the list.
+	 * @param record Preset record.
+	 * @param is_last True if preset list update operation can be considered concluded.
+	 */
+	void (*preset_update)(struct bt_has_client *client, uint8_t index_prev,
+			      const struct bt_has_preset_record *record, bool is_last);
+
+	/**
+	 * @brief Callback function for preset deletion notifications.
+	 *
+	 * The callback is called when the preset has been deleted by the remote server.
+	 *
+	 * @param client Client instance.
+	 * @param index Preset index.
+	 * @param is_last True if preset list update operation can be considered concluded.
+	 */
+	void (*preset_deleted)(struct bt_has_client *client, uint8_t index, bool is_last);
+
+	/**
+	 * @brief Callback function for preset availability notifications.
+	 *
+	 * The callback is called when the preset availability change is notified by the remote
+	 * server.
+	 *
+	 * @param client Client instance.
+	 * @param index Preset index.
+	 * @param available True if available, false otherwise.
+	 * @param is_last True if preset list update operation can be considered concluded.
+	 */
+	void (*preset_availability)(struct bt_has_client *client, uint8_t index, bool available,
+				    bool is_last);
+
+	/**
+	 * @brief Callback function for features notifications.
+	 *
+	 * The callback is called when the Hearing Aid Features change is notified by the remote
+	 * server.
+	 *
+	 * @param client Client instance.
+	 * @param features Updated features value.
+	 */
+	void (*features)(struct bt_has_client *client, enum bt_has_features features);
+
+	/**
+	 * @brief Callback function for command operations.
+	 *
+	 * The callback is called when command status is received from the remote server.
+	 *
+	 * @param client Client instance.
+	 * @param err 0 on success, ATT error otherwise.
+	 */
+	void (*cmd_status)(struct bt_has_client *client, uint8_t err);
+};
+
+/**
+ * @brief Initialize Hearing Access Service Client.
+ *
+ * Register Hearing Access Service Client callbacks.
+ *
+ * @return Zero on success or (negative) error code on failure.
+ */
+int bt_has_client_init(const struct bt_has_client_cb *cb);
+
+/**
+ * @brief Bind service.
+ *
+ * Initiate service binding to remote device implementing HAS Server role.
+ *
+ * @param conn Bluetooth connection object.
+ * @param[out] client Valid client instance on success.
+ *
+ * @return Zero on success or (negative) error code on failure.
+ */
+int bt_has_client_bind(struct bt_conn *conn, struct bt_has_client **client);
+
+/**
+ * @brief Unbind service.
+ *
+ * Unbind an service service or cancel pending service binding.
+ * The
+ *
+ * @param client Client instance.
+ *
+ * @return Zero on success or (negative) error code on failure.
+ */
+int bt_has_client_unbind(struct bt_has_client *client);
+
+/**
+ * @brief Read Preset Records command.
+ *
+ * Client method to read up to @p max_count presets starting from given @p index.
+ * The status is returned in the @ref bt_has_client_cb.cmd_status callback.
+ * The preset records are returned in the @ref bt_has_client_cb.preset_read_rsp callback
+ * (called once for each preset).
+ *
+ * @param client Client instance.
+ * @param index The index to start with.
+ * @param max_count Maximum number of presets to read.
+ *
+ * @return 0 in case of success or negative value in case of error.
+ */
+int bt_has_client_cmd_presets_read(struct bt_has_client *client, uint8_t index, uint8_t max_count);
+
+/**
+ * @brief Set Active Preset command.
+ *
+ * Client procedure to set preset identified by @p index as active.
+ * The status is returned in the @ref bt_has_client_cb.cmd_status callback.
+ *
+ * @param client Client instance.
+ * @param index Preset index to activate.
+ * @param sync Request active preset synchronization in set.
+ *
+ * @return 0 in case of success or negative value in case of error.
+ */
+int bt_has_client_cmd_preset_set(struct bt_has_client *client, uint8_t index, bool sync);
+
+/**
+ * @brief Activate Next Preset command.
+ *
+ * Client procedure to set next available preset as active.
+ * The status is returned in the @ref bt_has_client_cb.cmd_status callback.
+ *
+ * @param client Client instance.
+ * @param sync Request active preset synchronization in set.
+ *
+ * @return 0 in case of success or negative value in case of error.
+ */
+int bt_has_client_cmd_preset_next(struct bt_has_client *client, bool sync);
+
+/**
+ * @brief Activate Previous Preset command.
+ *
+ * Client procedure to set previous available preset as active.
+ * The status is returned in the @ref bt_has_client_cb.cmd_status callback.
+ *
+ * @param client Client instance.
+ * @param sync Request active preset synchronization in set.
+ *
+ * @return 0 in case of success or negative value in case of error.
+ */
+int bt_has_client_cmd_preset_prev(struct bt_has_client *client, bool sync);
+
+/**
+ * @brief Write Preset Name command.
+ *
+ * Client procedure to change the preset name.
+ * The status is returned in the @ref bt_has_client_cb.cmd_status callback.
+ *
+ * @param client Client instance.
+ * @param index Preset index to change the name of.
+ * @param name Preset name to write.
+ *
+ * @return 0 in case of success or negative value in case of error.
+ */
+int bt_has_client_cmd_preset_write(struct bt_has_client *client, uint8_t index, const char *name);
+
+/** Hearing Aid device Info Structure */
+struct bt_has_client_info {
+	/** Local identity */
+	uint8_t id;
+	/** Peer device address */
+	const bt_addr_le_t *addr;
+	/** Hearing Aid Type. */
+	enum bt_has_hearing_aid_type type;
+	/** Hearing Aid Features. */
+	enum bt_has_features features;
+	/** Hearing Aid Capabilities. */
+	enum bt_has_capabilities caps;
+	/** Active Preset Index.
+	 *
+	 *  Only applicable if @ref BT_HAS_PRESET_SUPPORT is set in @ref caps.
+	 */
+	uint8_t active_index;
+};
+
+/**
+ * @brief Get information for the remote Hearing Aid device.
+ *
+ * @param client Client instance.
+ * @param[out] info Info object.
+ *
+ * @return Zero on success or (negative) error code on failure.
+ */
+int bt_has_client_info_get(const struct bt_has_client *client, struct bt_has_client_info *info);
+
+/**
+ * @brief Get the connection pointer of a HAS Client instance
+ *
+ * Get the Bluetooth connection pointer of a HAS Client instance.
+ *
+ * @param client Client instance.
+ * @param[out] conn Connection pointer.
+ *
+ * @return 0 if success, errno on failure.
+ */
+int bt_has_client_conn_get(const struct bt_has_client *client, struct bt_conn **conn);
