+++ tests/bluetooth/shell/testcase.yaml
@@ -258,11 +258,12 @@ tests:
     build_only: true
     extra_configs:
       - CONFIG_BT_HAS=n
-  bluetooth.audio_shell.no_has_client:
+  bluetooth.audio_shell.no_hap_harc:
     extra_args: CONF_FILE="audio.conf"
     build_only: true
     extra_configs:
       - CONFIG_BT_HAS_CLIENT=n
+      - CONFIG_BT_HAP_HARC=n
   bluetooth.shell.audio.no_tbs:
     extra_args: CONF_FILE="audio.conf"
     build_only: true
